+++ ru.taximaxim.codekeeper.core.tests/src/main/java/ru/taximaxim/codekeeper/core/PgDiffTest.java
@@ -602,7 +602,13 @@ class PgDiffTest {
             "compare_pg_unicode_identifiers",
             // Test scenario when one child partition index is dropped
             "drop_pg_child_partition_index",
-            "add_objects_with_same_name"
+            "add_objects_with_same_name",
+            "drop_pg_child_partition_index",
+            // Test scenario when trigger is enabled on child partition table and disabled on parent
+            "modify_partition_table_trigger_enabled",
+            // Test scenario when triggers are removed from parition table
+            "modify_partition_table_removed_triggers"
+
     })
     void runDiff(String fileNameTemplate) throws IOException, InterruptedException {
         TestUtils.runDiff(fileNameTemplate, DatabaseType.PG, PgDiffTest.class);