+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/modify_partition_table_trigger_enabled_diff.sql
@@ -1,0 +1,13 @@
+SET search_path = pg_catalog;
+
+ALTER TABLE public.order_book DISABLE TRIGGER swt_pk_order_book1;
+
+ALTER TABLE public.order_book DISABLE TRIGGER swt_pk_order_book2;
+
+ALTER TABLE public.order_book_p1703 ENABLE TRIGGER swt_pk_order_book2;
+
+ALTER TABLE public.order_book_p1703 ENABLE TRIGGER swt_pk_order_book1;
+
+ALTER TABLE public.order_book_p1704 ENABLE TRIGGER swt_pk_order_book1;
+
+ALTER TABLE public.order_book_pmax ENABLE TRIGGER swt_pk_order_book2;